# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SimpleShop API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Created
  /api/auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
  /api/auth/logout:
    post:
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/products:
    get:
      summary: List products
      responses:
        '200':
          description: OK
  /api/products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/categories:
    get:
      summary: List categories
      responses:
        '200':
          description: OK
  /api/products/{id}/reviews:
    get:
      summary: List reviews by product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      summary: Add review
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating, comment]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        '201':
          description: Created
  /api/products/{id}/reviews/{reviewId}:
    delete:
      summary: Delete review
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/orders:
    get:
      summary: List orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Create order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items, total]
              properties:
                items:
                  type: array
                  items:
                    type: object
                total:
                  type: number
      responses:
        '201':
          description: Created
  /api/orders/{id}:
    get:
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT